Refactored code as a code review
