+++ lib/app_colors.dart
@@ -1,0 +1,5 @@
+import 'package:flutter/material.dart';
+
+const primaryGreenColor = Color(0xff438E64);
+final whiteWithOpacity = Colors.white.withOpacity(0.8);
+const deepGreenAccent = Color(0xff327250);
